@@ -10,9 +10,9 @@
         "guzzlehttp/guzzle": "^7.2",
         "justinrainbow/json-schema": "^5.2",
         "laminas/laminas-feed": "^2.22",
-        "laravel/framework": "^10.10",
+        "laravel/framework": "^11.0",
         "laravel/horizon": "^5.26",
-        "laravel/sanctum": "^3.3",
+        "laravel/sanctum": "^4.0",
         "laravel/tinker": "^2.8",
         "marcandreappel/laravel-backblaze-b2": "^2.0",
         "railsware/mailtrap-php": "^1.9",
@@ -28,7 +28,7 @@
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
-        "nunomaduro/collision": "^7.0",
+        "nunomaduro/collision": "^8.1",
         "pestphp/pest": "^2.29",
         "pestphp/pest-plugin-laravel": "^2.2",
         "phpunit/phpunit": "^10.1",
